seperated functions to files Fixes #387
